Extract CodeInclusion module into separate file
